refactor(tui): modularize TUI into package structure
Reorganizes TUI code into modular structure:
- tui/base.py: BaseBrowserApp, DetailView, COMMON_CSS, helpers
- tui/widgets.py: SearchInput, HelpScreen, JumpDialog, SaveDialog
- tui/catalog/: CatalogBrowserApp, CatalogDetailView, ActionMenu

Benefits:
- Clear separation of concerns
- Easier to add new TUI types (diff, view)
- DRY shared components
- Better testability

Old imports (tui_base.py, catalog_tui.py) re-export from
new locations for backwards compatibility.

Authored by: Aaron Lippold<lippold@gmail.com>
