+++ src/cis_bench/cli/commands/tui/catalog/__init__.py
@@ -1,0 +1,12 @@
+"""Catalog browser TUI components."""
+
+from cis_bench.cli.commands.tui.catalog.actions import ActionMenu
+from cis_bench.cli.commands.tui.catalog.app import CatalogBrowserApp, run_catalog_browser
+from cis_bench.cli.commands.tui.catalog.detail import CatalogDetailView
+
+__all__ = [
+    "ActionMenu",
+    "CatalogBrowserApp",
+    "CatalogDetailView",
+    "run_catalog_browser",
+]
